# specs/005-field-intelligence-safety/contracts/time-tracking.openapi.yaml
openapi: 3.0.3
info:
  title: Time Tracking API
  version: 1.0.0
  description: GPS-based time tracking with auto-detection for Feature 005
paths:
...
  /api/time/clock-out:
    post:
      summary: Clock out with GPS
      operationId: clockOut
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [location]
              properties:
                location:
                  $ref: '#/components/schemas/Location'
                suggested_time:
                  type: string
                  format: date-time
                  description: Used for forgot-to-clock-out scenarios
      responses:
        '200':
          description: Clocked out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeEntryResponse'
...
components:
  schemas:
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        accuracy:
          type: number
          format: float
    TimeEntryResponse:
      type: object
      properties:
        time_entry_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [job_work, travel, break, admin, equipment_swap]
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        start_location:
          $ref: '#/components/schemas/Location'
        end_location:
          $ref: '#/components/schemas/Location'
        auto_created:
          type: boolean
        flagged_for_review:
          type: boolean
        flag_reason:
          type: string
...